Consolidate Attendance Mail Callbacks, DRYness
Moved sending of Attendance registration mail from within the TT controller into
the `queue_mais` method on the after_create hook of the Attendance. Eventually
we're likely to want to move all this behaviour out into service objects, but
it makes it easy to track down the source of mails for now.