Clear ActiveMailer deliveries after all specs
ActiveMailer is set to clear the deliveries array after Mailer specs run,
but not after request or controller specs. This causes the array to be
prepopulated when specs run after a request or controller array that
creates mail. This fixes that.

Also switches to using lets in the attendance mailer spec for increased
clarity