Point to missing-city-image in the most convoluted way possible.
In production assets are precompiled and asset_host is set to the Cloudfront URL,
so we need to point to the precompiled path with the digest. The Rails helper
generates a URL that points to the cloudfront host (because of asset_host config),
so we do a manual filesystem lookup here for a file that starts with missing-city-image,
and if it exists, generate a /assets/ url from that. If it doesn't, fallback to just
the name of the file (this will be the case in development env when asset_host
is nil and we're serving from asset pipeline