Improve logic for create/update button text.
These buttons are identical except the CSS class and the button text.
Since I don't know if the classes are important and since I don't
want a grid level class to dictate hiding/showing an unrelated DOM element,
I am changing to using a pattern that is used elsewhere in the app
to determine if this is an edit or new form: the .persisted? method in the modl.

A better solution would be to have separate forms for new and edit
or to set a virtual attribute in the controller. But this is better
than the current solution and matches the pattern in other places