Test passed intermittently; wrong object passed in
`mark_attendance_mailer` expects a Tea Time ID, not an Attendance object. The
test was being passed an Attendance ID.  This just so happened to work most of
the time because there are TTs littering the DB from prior tests. This does not
work when mailer specs are run earlier, and so it failed intermittently under
CI.