Add feature to sync users with SendGrid contact database
Adds a rake task, sendgrid_list:sync, to perform a one-way sync.

Adds create/update/delete hooks to the User model to update the user
to SendGrid when it changes.

Adds create/update hooks to the Attendance model to update the user's
tea time attendance data to SendGrid when it changes.