Update sendgrid_list:sync script to remove non-existent users
The sync script will now load all users from SendGrid's contacts DB
and remove users that do not exist in the user database. This keeps
the SendGrid contacts DB consistent with the user database, such as
in cases where the ActiveRecord callbacks in the application which
are supposed to sync users state between TWS and SendGrid somehow
miss updates.